requirements.txt
boto3==1.9.196
botocore==1.12.196
certifi==2019.6.16
chardet==3.0.4
docutils==0.14
future==0.17.1
idna==2.8
jmespath==0.9.4
oauthlib==3.0.2
Pillow==6.1.0
pkg-resources==0.0.0
pymongo==3.8.0
python-dateutil==2.8.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
